Update Mentorship model and handlers to use mentorship_id instead of config_id

+++ src/functions/Mentorship/delete/handler.ts
@@ -7,6 +7,8 @@ import { i18nMiddleware } from "../../../middlewares/i18n";
 import { schemaValidator } from "../../../libs/lambda";
 import { object, string } from "yup";
 import i18n from "../../../libs/i18n";
+import { cognitoMiddleware } from "../../../middlewares/JWT";
+import { Roles } from "../../../utils/enums";
 
 const i18nString = (key: string) => i18n.t("Mentorship.delete." + key);
 
@@ -16,18 +18,36 @@ const handler = async (
   try {
     const prisma = initializePrisma();
     const { mentorship_id } = event.pathParameters ?? {};
+    const { user_id = "" } = event.headers;
 
-    await prisma.appointment.deleteMany({
-      where: {
-        mentorship_id: Number(mentorship_id),
-      },
-    });
-
-    await prisma.mentorship.delete({
+    const mentorship = await prisma.mentorship.findUnique({
       where: {
         id: Number(mentorship_id),
+        tutor_id: Number(user_id),
       },
     });
+
+    if (!mentorship) {
+      return Responses._404({
+        message: i18nString("validations.mentorshipNotFound"),
+      });
+    }
+
+    const deleteAppointments = prisma.appointment.deleteMany({
+      where: {
+        mentorship_id: Number(mentorship_id),
+        tutor_id: Number(user_id),
+      },
+    });
+
+    const deleteMentorship = prisma.mentorship.delete({
+      where: {
+        id: Number(mentorship_id),
+        tutor_id: Number(user_id),
+      },
+    });
+
+    await prisma.$transaction([deleteAppointments, deleteMentorship]);
 
     return Responses._200({ message: i18nString("success") });
   } catch (error) {
@@ -39,6 +59,7 @@ const handler = async (
 export const deleteMentorship = middy(handler).use([
   jsonBodyParser(),
   i18nMiddleware(),
+  cognitoMiddleware([Roles.Tutor]),
   schemaValidator({
     pathParameters: object({
       mentorship_id: string().required(
